Created rmNode, rmNodeAttributes, findChildWithID.
The functions that were created are designed to remove a node from and ovf
or remove attributes from a given node within an ovf.

rmNode: Removes a given node to.

rmNodeAttributes: This function will remove the attributes for a given section.
    Passing False to strict will mean that no errors will be thrown
    if an attribute for the given node is not found.

findChildWithID: This helper function will find a child node with the
    given id.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>