Removed findChildWithID and modified getSection.
The method getSection was doing what findChildwithID was doing. So in an effort
to consolidate the two functions I removed findChildWithID and modified
getSection. The modification to get section was simply to check for different
id forms. For example the Disk section used ovf:diskId so I needed to check
for that.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>