replace Ovf.getSections with more generic getNodes and getNodesWithId
This replaces the getSections function from ovf/Ovf.py with a getNodes
function and getNodesWithId.  getNodes is a very generic function that
returns a list of elements in a node that are of a given tagname.  That
list can optionally be limited by a dictionary of attributes.

Ie, to find all 'VirtualSystemCollection' elements with an attributes
  ovf:id = "myId" and ovf:size = "10"
call
  getNodes(dom,'VirtualSystemCollection',{ 'ovf:id':'myId', 'ovf:size':10 })

The getNodesWithId(dom, tagName, id) is a shortcut method that essentiall
  does the following (with searching other 'id' attributes also)
  getNodesWithId(dom, tagName, { 'ovf:id': id } )

Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>