fix examples/sampleMkovf.sh by working around inadequacy in mkovf
Previously, examples/sampleMkovf.sh was failing to run with traceback like:
  File "./py/scripts/mkovf", line 601, in main
    commands[command]['func'](ovfFile, options)
  File "./py/scripts/mkovf", line 232, in createResources
    raise NotImplementedError, "The section with id, "+options.id+", was not found."
NotImplementedError: The section with id, MyLampService, was not found.

The issue is simply worked around here in the example script.  Previously
no sectionID (secID) was provided in the --virthw invocation.  The result
was that subsequent calls to --resource would fail as mkovf could not
find the VirtualHardwareSection to add the resource to.

We've got some general clean needed around how to specify from the command
line where to add nodes.  This is just a workaround for now.