require --init to mkovf to create new ovf
Before in main all the exceptions were being caught regardless of what they
were and the init function was being called. Also before you could just make
any sections and if the OVF did not exists the script would make it for you.
However now it will raise an error if the OVF does not exist and the user
does not use --init to create the new OVF.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
remove syntax errors in mkovf, make createFullOvfSample.sh execute successfully
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>