Changed the help section in mkovf.
The help section had new lines that were incorrect. This fixes those new
lines.

Also fixed typo in the help section for icon.

Lastly changed from ' to " and from having +" for a new line to "+.
This is to help make the code consistent.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
merged in previous non-posted cleanup patch from marcos
Signed-off-by: Scott Moser <smoser@us.ibm.com>