Implemented rmReferencesHandler function.
The function will remove a File from the References section of an ovf using
the flag --node. Added the flags --node and --strict to the common dictionary.
The flag node can be used to remove a node with a given id. The flag strict
will allow the user specify if they want to see errors such as when trying to
remove an attribute and the attribute does not exist. To remove the entire
References section the user can use the flag --section. Lastly remove ovfID
form being required because if an user wants to delete the entire section
then the id is not needed.

The change of name was done to make it more descriptive of what the function
actually does.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>