Re-worked the function rmNetworkHandler to use the changes in Ovf.
A few of the methods in Ovf.py that were being called from rmNetworkSection
had changed. Also the method changed to utilize --node and --strict flags.
The try catch block in the method is to verify the existence of a given network
with a given id.

Also in main had to change from not really showing the errors that could be
raised from creating an ovfFile object to actually showing them.

The change of name was done to make it more descriptive of what the function
actually does.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>