make xmllint --nonet work
Currently, xml.xsd is in the schemas dir, but xmllint will pull it
due to the "schemaLocation" value in ovf-envelope.xsd.
This makes 'xmllint --nonet' work.  Ie, the following will return
success:

xmllint --nonet --path schemas/ --noout \
   --schema schemas/ovf-envelope.xsd examples/someOVF.ovf

I find that generally a good thing, as it makes the schemas directory
"standalone" and allows tests to pass even with flaky or non-existent
network.

The negatives that I see to this right now are:
 a. I'm changing ovf-envelope.xsd .  It won't be bit-for-bit the same as
the official envelope
 b. We're caching xml.xsd, which could possibly get out of date (my
feeling is that it would be very bad for that URL to change its
contents)

Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>