Implemented rmDeploymentHandler.
This function will remove the Deployment Options Section OR remove a
configuration from an existing Deployment Options Section. The user can also
remove the Label or Description from the given configuration. To remove the
DeploymentOptionsSection use --section.

In the commands section updated the rmDeploymentHandler and rmNetworkHandler
to match the correct function name with the correct function. Also removed
config from the deploy commands because the user can use node to remove the
given configuration with a specific id.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>