Updated rmReferencesHandler to raise ValueError appropiately.
In rmReferencesHandler removed the check if it was strict after looking for a
specific File. The reason for removing that is that if the file with a given
id does not exist then it will throw a IndexError, which is not very
descriptive of why it happened.

Removed the file id as being required. If the user wants to remove the entire
section there is no need to specify a file id.

Also updated the epydoc.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>