Updated rmDiskHandler to raise ValueError appropiately.
Just like the rmReferencesHandler took out strict after looking for a specific
Disk. The reason for removing that was the same as stated above. Also had to
add a way to remove the Info within that section and raise an error if the Info
was not present and was flagged to be removed. To find an Info section the user
can specify the infoID which is the ovf:msgid for that Info. Alternatively the
user can remove the first instance of Info that is found in the section.

Updated the help section for the commands. Also made infoID store a value
instead than a Boolean.

Also updated the epydoc.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
spelling in commit msg
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>