Updated rmNetworkHandler to raise ValueError appropiately.
In the NetworkSection had to re-work how the Info node was being removed if it
is flagged for removal. Before it was getting the node and trying to remove it
in one line. However if the node did not exist it would throw an IndexError.
Also took out the try-catch within the method in order to make it more
consistent with how the other methods raise an exception. Also had to add a way
to remove the Info within that section and raise an error if the Info was not
present and was flagged to be removed. To find an Info section the user can
specify the infoID which is the ovf:msgid for that Info. Alternatively the user
can remove the first instance of Info that is found in the section. To remove a
Description from a network the user can give the msgid for that Description or
can give the flag for removing a description without an id. Note that the
latter will remove all the Descriptions from a given Network.

Updated the help section for the commands to match what each flag was doing.
Also added descID which will store the id for a given Description.

Also updated the epydoc.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
converted tab indentation to space, spelling in commit msg
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>