Updated rmDeploymentHandler to to raise ValueError appropiately.
For the rmDeploymentHandler function added checking to ensure that if a node or
node attribute is to be deleted that it must be present otherwise a ValueError
is thrown. Also changed how the functions find the Label and Description for
a given configuration. Note that the way the Info, Description, Label are
acquired are very similar so they will be merged into a generic method.

Updated the commands for the deployment so that the flags that get ids are not
Boolean.

Also updated the epydoc.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
spelling in commit msg
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>