fix logic error in OvfSetTestCase.py:test_verifyManifest
The loop was intending to go through a list of files and set a bogus checksum
for given href (Ubuntu1.vmdk), and remove other nodes.  Instead, of removing
the current node it would always remove the first.  This would lead to key
error later as Ubuntu1.vmdk was getting removed accidentally.