clean up getReferencedFilesFromOvf
Change prototype of getReferencedFilesFromOvf from taking a fileName
to taking an envelope.  This way you can pass it an envelope and avoid it
needing to parseFile

getReferencedFilesFromOvf had some problems where if a '<File>' element
did not have a href attribute (or fileid attribute) it would throw an
error.  This change fixes that bug and shortens the method.

Additionally, the routine previously had function that ensured the list
was in ".ova sorted order".  Ie, the .mf and .cert files would be first if
present.  I've removed that code.  Which at the moment makes Unit Test
fail.

Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>