modify OvfSetTextCase:test_writeAsTar to properly test order of .mf, .cert
test_writeAsTar had a logic flaw in its test.  It was asserting that
if the <File> elements had a .mf or .cert file in them, then they
had to be in the correct position in the tar archive.  The logic would
then depend on the present of the .mf or .cert as a <File> element.
That is not likely to happen.

Second change here is that the spec now allows for manifest and cert to
be at the end of the archive (last 2 files or last file).

We're actually only passing this test by chance right now.  There are still
2 things to do:
 1. remove the .mf and .cert files from the .mf in ourOVF.ovf example ovf
    (and clean up the fallout)
 2. make OvfSet::writeAsTar (and OvfSet::writeAsDir) actually write .mf
    and .cert file if instructed to do so

Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>