remove .mf and .cert file from ourOvf.ovf References section
This is a fairly straightforward removal of the .cert and .mf files from
the <References> section of ourOvf.ovf in py/tests/test_files.

While the spec is not 100% clear, we do not expect that it intends for the
files to be present here.  The modification to the spec to allow .mf and
.cert files to be the last files in a .ova really points to the intent
that these 2 files could be generated and appended to an already existing
.ova.  That flow doesn't really work if the .mf and .cert files have to be
listed in the .ovf file

commit 283e8ec11c7592ab7ff32efa8685c6795973672b mentions two TODOs. This
addresses the first (remove the .mf and .cert from ourOVF.ovf).  The second
(correctly write .mf and .cert is still TODO)

Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>