Created createTextDescriptionOfNodeList.
This function will get information from a list of nodes and return a list
containing the information found for each node. This function is called in the
class MultipleNodeError to create the error message that the class shows. For
all sections the method uses Info as a descriptor of an individual node. For
oter nodes it uses the attributes contained within the node. In the event that
the list of nodes passed in is either rasd or vssd elements, it will use the
ElementName. This function uses the updated method getNodes in Ovf.

Signed-off-by: Marcos Cintron <macintro@us.ibm.com>
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>