Add ova --pack and --unpack
New version of this patch addressing the comments from Scott.  The changes
in OvfSet writeAsDir caused a test case to fail.  The test case for
path = None passed is no longer valid.  Since path is now defaulted to
ovfFile.path when None is passed, this will no longer result in an
exception.  I changed the test case to instead verify that an IOError is
raised when a non-existent path is passed.

Signed-off-by: David L. Leskovec <dlesko@linux.vnet.ibm.com>
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>