Library to create and manage ovf-env.xml file.
Adding three files that make library for creating and managing ovf-env.xml
files. EnvironmentSection.py file is the main class for creating/managing
environment file. PlatformSection.py is used for PlatformSection which is
part of the environment. Constants.py is the file storing all the
constants used in this library.

Signed-off-by: Sharad Mishra <snmishra@us.ibm.com>
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>