Enhance getDefaultConfiguration
This patch changes Ovf.getDefaultConfiguration to first find a
configuration element with the attribute ovf:default == true.  If found,
this is the default configuration.  If no configuration has this
attribute, then the previous behavior is followed and the first
Configuration node is considered the default.  If no Configuration
elements are specifed, then None is returned.

Signed-off-by: David L. Leskovec <dlesko@linux.vnet.ibm.com>
Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>