Enable virtual system network XML
Hi,

This patch fills in some and enables the network support for a virtual
system within a virtual appliance.  A network device is provided to the virtual
machine.  This network device is attached to the libvirt network named by
the <rasd:Connection> element.  So, an Item block in the VirtualHardwareSection
like this:

<Item>
    <rasd:AutomaticAllocation>true</rasd:AutomaticAllocation>
    <rasd:Caption>Ethernet adapter on 'default'</rasd:Caption>
    <rasd:Connection>default</rasd:Connection>
    <rasd:Description>Ethernet Adapter</rasd:Description>
    <rasd:ElementName>Network Adapter</rasd:ElementName>
    <rasd:InstanceID>3</rasd:InstanceID>
    <rasd:ResourceType>10</rasd:ResourceType>
</Item>

will result in a network device attached to the libvirt 'default' network
(which is provided 'out of the box' by libivirt).

Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>