Add xen pv support
Hi,

This patch adds support for virtual systems deployed as Xen
paravirtualized guests.  Fullyvirtualized guests require another patch to
fix the way the domain type is selected.  I'll post that patch in a day or
so.  This patch applies on top of mainline plus the network patch I
posted back on the 10th.

Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>