Enable Xen full virtualization
Hi,

This patch enables virutal systems implemetented as Xen fully
virtualized guests.  These changes were required to enable this:
1) Virtual system type == xenfv for fully virutalized guest
2) Virtual system type == xenpv for para-virutalized guest
3) -c, --connect parameter removed from ova --runtime.  The libvirt
connection string is now generated based on the virtual system type.
4) -v, --virt parameter added to ova --runtime.  This can be used to
specify the virtual system type for all virtual systems in the appliance.
This value will override a valued specified in the ovf file.  If this
parameter is not specified, the type for each virtual system is taken
from the ovf file.