if TMPDIR environment is set, use it for mkdtemp
We changed the target directory where .ova temporary extraction was
done by calling mkdtemp with an explicit path to the location of the
.ova file.

A call to mkdtemp with dir=None will use TMPDIR if set in the
environment.  This change lets the user specify TMPDIR.  If it is not
set in the environment, then use the path above.

Signed-off-by: Scott Moser <ssmoser@linux.vnet.ibm.com>